requirements.txt
Flask>=1.1
face_recognition
opencv-python
numpy
Pillow
